CHANGES: 1) fixed major bug in discretize_in_space() function member - lss_pde_utility.h 2) ran all the unit tests 3) added additional unit test for explicit CUDA solvers testing black-scholes equation
TODO:
1) add other tests on different extensions of BS model
2) start thinking about 2 dim heat equation and underlying system
solvers

+++ pde_solvers/one_dim/variable_coefficients/lss_space_variable_heat_explicit_schemes_cuda_policy.cu
@@ -1005,7 +1005,7 @@ void heat_euler_scheme_backward_policy<
   auto const &left = dirichlet_boundary.first;
   auto const &right = dirichlet_boundary.second;
 
-  float time = terminal_time - k;
+  double time = terminal_time - k;
 
   // prepare pointers for PDE space variable coeffs on device:
   double *d_A = NULL;
